+++ packages/adapter-cloudflare/files/worker.js
@@ -45,7 +45,7 @@ export default {
 
 		// dynamically-generated pages
 		try {
-			return await app.render(req);
+			return await app.render(req, { platform: { env } });
 		} catch (e) {
 			return new Response('Error rendering route: ' + (e.message || e.toString()), { status: 500 });
 		}