+++ packages/adapter-netlify/src/handler.d.ts
@@ -1,3 +1,3 @@
-declare module 'APP' {
+declare module '0APP' {
 	export { App } from '@sveltejs/kit';
 }